+++ ngyn_resource_planning/models/__init__.py
@@ -1,3 +1,4 @@
 from . import project_task
 from . import hr_employee
 from . import resource_assignment
+from . import account_analytic_line